fix: use decimal instead of double for StatementRequest.Amount

+++ PennyPincher.Contracts/Statements/StatementRequest.cs
@@ -3,7 +3,7 @@
 public record StatementRequest(
         DateTime Date,
         int AccountId,
-        double Amount,
+        decimal Amount,
         string Description,
         int CategoryId
     );
